Move V2 CodeActions models to separate folder and namespace
