+++ src/OmniSharp.Abstractions/Models/v2/CodeActions/CodeAction.cs
@@ -1,4 +1,4 @@
-namespace OmniSharp.Models.V2
+namespace OmniSharp.Models.V2.CodeActions
 {
     public class OmniSharpCodeAction
     {
