+++ src/OmniSharp.Abstractions/Models/v2/CodeActions/GetCodeActionsRequest.cs
@@ -1,6 +1,6 @@
 using OmniSharp.Mef;
 
-namespace OmniSharp.Models.V2
+namespace OmniSharp.Models.V2.CodeActions
 {
     [OmniSharpEndpoint(OmniSharpEndpoints.V2.GetCodeActions, typeof(GetCodeActionsRequest), typeof(GetCodeActionsResponse))]
     public class GetCodeActionsRequest : Request, ICodeActionRequest
