+++ src/OmniSharp.Abstractions/Models/v2/CodeActions/RunCodeActionRequest.cs
@@ -1,6 +1,6 @@
 using OmniSharp.Mef;
 
-namespace OmniSharp.Models.V2
+namespace OmniSharp.Models.V2.CodeActions
 {
     [OmniSharpEndpoint(OmniSharpEndpoints.V2.RunCodeAction, typeof(RunCodeActionRequest), typeof(RunCodeActionResponse))]
     public class RunCodeActionRequest : Request, ICodeActionRequest
